refactor: execution(TaskRun, JobRun, WorkflowRun) 도메인 설정
