+++ apps/user-service/src/main/java/site/icebang/domain/execution/mapper/WorkflowRunMapper.java
@@ -1,8 +1,8 @@
-package site.icebang.domain.workflow.mapper.execution;
+package site.icebang.domain.execution.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
 
-import site.icebang.domain.workflow.model.execution.WorkflowRun;
+import site.icebang.domain.execution.model.WorkflowRun;
 
 @Mapper
 public interface WorkflowRunMapper {
