+++ apps/user-service/src/main/java/site/icebang/domain/execution/model/WorkflowRun.java
@@ -1,4 +1,4 @@
-package site.icebang.domain.workflow.model.execution;
+package site.icebang.domain.execution.model;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
