+++ apps/user-service/src/main/java/site/icebang/domain/workflow/service/WorkflowExecutionService.java
@@ -10,14 +10,14 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import site.icebang.domain.workflow.mapper.JobMapper;
-import site.icebang.domain.workflow.mapper.execution.JobRunMapper;
-import site.icebang.domain.workflow.mapper.execution.TaskRunMapper;
-import site.icebang.domain.workflow.mapper.execution.WorkflowRunMapper;
+import site.icebang.domain.execution.mapper.JobRunMapper;
+import site.icebang.domain.execution.mapper.TaskRunMapper;
+import site.icebang.domain.execution.mapper.WorkflowRunMapper;
 import site.icebang.domain.workflow.model.Job;
 import site.icebang.domain.workflow.model.Task;
-import site.icebang.domain.workflow.model.execution.JobRun;
-import site.icebang.domain.workflow.model.execution.TaskRun;
-import site.icebang.domain.workflow.model.execution.WorkflowRun;
+import site.icebang.domain.execution.model.JobRun;
+import site.icebang.domain.execution.model.TaskRun;
+import site.icebang.domain.execution.model.WorkflowRun;
 import site.icebang.domain.workflow.runner.TaskRunner;
 
 @Slf4j
