+++ apps/user-service/src/main/resources/mybatis/mapper/JobRunMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="site.icebang.domain.workflow.mapper.execution.JobRunMapper">
+<mapper namespace="site.icebang.domain.execution.mapper.JobRunMapper">
 
-    <resultMap id="JobRunResultMap" type="site.icebang.domain.workflow.model.execution.JobRun">
+    <resultMap id="JobRunResultMap" type="site.icebang.domain.execution.model.JobRun">
         <id property="id" column="id"/>
         <result property="workflowRunId" column="workflow_run_id"/>
         <result property="jobId" column="job_id"/>
