+++ apps/user-service/src/main/resources/mybatis/mapper/WorkflowRunMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="site.icebang.domain.workflow.mapper.execution.WorkflowRunMapper">
+<mapper namespace="site.icebang.domain.execution.mapper.WorkflowRunMapper">
 
-    <resultMap id="WorkflowRunResultMap" type="site.icebang.domain.workflow.model.execution.WorkflowRun">
+    <resultMap id="WorkflowRunResultMap" type="site.icebang.domain.execution.model.WorkflowRun">
         <id property="id" column="id"/>
         <result property="workflowId" column="workflow_id"/>
         <result property="traceId" column="trace_id"/>
